pyproject.toml
[tool.poetry]
name = "drf-search-service"
version = "0.1.0"
description = "Django REST Framework Search Service"
authors = ["Finlay McCourt <finlaymccourt@gmail.com>"]
readme = "README.md"
packages = [
    { include = "search_service" },
]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.10.0"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
requests = "^2.27.1"
bleach = "^4.1.0"
django-extensions = "^3.1.5"
django-model-utils = "^4.2.0"
ordered-set = "^4.0.2"
psycopg2-binary = "^2.9.3"
django-filter = "^21.1"
python-dateutil = "^2.8.2"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
